+++ R/polmineR.R
@@ -18,6 +18,9 @@ NULL
 
 setOldClass("htmlwidget")
 
+#' @exportMethod zoom
+#' @noRd
+setGeneric("zoom", function(x, ...) standardGeneric("zoom"))
 
 
 #' polmineR-package